Install enzyme-adapter-react-16 module. Configure Enzyme adpater in each Component test suit. Rename flashMessages to flashMessagesActions. Add tempPolyfil.js config file.
