+++ client/src/__test__/tempPolyfill.js
@@ -1,0 +1,5 @@
+const raf = global.requestAnimationFrame = (cb) => {
+  setTimeout(cb, 0);
+}
+
+export default raf
