+++ client/src/components/Login/Subdomain.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { Validation } from '../../utils'
-import { subdomainRequest } from '../../actions/authenticationAction'
-import { addFlashMessage } from '../../actions/flashMessages'
+import { subdomainRequest } from '../../actions/authenticationActions'
+import { addFlashMessage } from '../../actions/flashMessagesActions'
 import classnames from 'classnames'
 
 // Localization
