+++ client/src/components/Signup/Page.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import Form from './Form'
-import { signupRequest, isAccountExists, isUserExists} from '../../actions/authenticationAction'
-import { addFlashMessage } from '../../actions/flashMessages'
+import { signupRequest, isAccountExists, isUserExists} from '../../actions/authenticationActions'
+import { addFlashMessage } from '../../actions/flashMessagesActions'
 import FlashMessagesList from '../../flash/FlashMessagesList'
 
 // Localization
